+++ changelog.md
@@ -17,6 +17,7 @@ __[changes]__
 __[maintenance]__
 
 * NLog library is updated from 5.2.5 to 5.2.7.
+* Update certificate information for IrfanView installer.
 
 ## Version 2023.11.14.0